Update "VM acceleration" link in _android-setup.md
"Configure hardware acceleration for the Android Emulator" is a long page. If the anchor text is accurate and only VM acceleration is required, then it should go straight to the "Configuring VM acceleration"

Top-of-the-page link caused me to absent-mindedly start working through the whole page, starting with "Configure graphics acceleration". It seems configuring graphics acceleration is not a prereq for just configuring VM acceleration though (correct me if I'm wrong), so we should help readers skip to the point.